:bug: Extending fix to all action that create a prompt so all prompts are aligned, small refactor
