+++ src/main/kotlin/com/github/sloppylopez/moneypennyideaplugin/intentions/RefactorIntentionFactory.kt
@@ -34,7 +34,7 @@ class RefactorIntentionFactory(private val project: Project) {
                 override fun isAvailable(project: Project, editor: Editor?, file: PsiFile?): Boolean = true
 
                 override fun invoke(project: Project, editor: Editor?, file: PsiFile?) {
-                    promptPanelFactory.sendToContentPrompt(editor, service.psiFileToFile(file!!))
+                    promptPanelFactory.sendToContentPrompt2(editor, service.psiFileToFile(file!!))
                 }
 
                 override fun startInWriteAction(): Boolean = false
