+++ src/main/kotlin/com/github/sloppylopez/moneypennyideaplugin/toolWindow/TextAreaFactory.kt
@@ -1,14 +1,16 @@
 package com.github.sloppylopez.moneypennyideaplugin.toolWindow
+
 import com.github.sloppylopez.moneypennyideaplugin.components.MoneyPennyTextArea
 import com.intellij.notification.*
 import com.intellij.openapi.components.Service
 import com.intellij.ui.JBColor
+import com.intellij.util.ui.JBUI
 import javax.swing.BorderFactory
 import javax.swing.JTextArea
 
 @Service(Service.Level.PROJECT)
 class TextAreaFactory {
-    fun createTextArea(text: String, rows: Int, columns: Int, imageBackground: String?): JTextArea {
+    private fun createTextArea(text: String, rows: Int, columns: Int, imageBackground: String?): JTextArea {
         return MoneyPennyTextArea(imageBackground).apply {
             this.text = text
             lineWrap = true
@@ -17,5 +19,18 @@ class TextAreaFactory {
             this.columns = columns
             border = BorderFactory.createLineBorder(JBColor.LIGHT_GRAY) // Set red border
         }
+    }
+
+    fun createPaddedTextArea(
+        text: String,
+        rows: Int,
+        columns: Int,
+        imageBackground: String? = null,
+        textArea: JTextArea? = this.createTextArea(text, rows, columns, imageBackground)
+    ): JTextArea {
+        val padding = JBUI.insets(5) // Adjust the padding values as needed
+        val border = BorderFactory.createEmptyBorder(padding.top, padding.left, padding.bottom, padding.right)
+        textArea?.border = border
+        return textArea!!
     }
 }
